Avance 4

+++ dataBaseManager/src/main/java/edu/upvictoria/fpoo/InterpreteSQL.java
@@ -9,7 +9,7 @@ public class InterpreteSQL {
 
     private String rutaTrabajo;
     private String sentenciaSql;
-
+    BufferedReader bufer = new BufferedReader(new InputStreamReader(System.in));
 
 
 //--------------COMANDOS
@@ -23,7 +23,7 @@ public class InterpreteSQL {
                 use(rutaNueva);
 
             }else {
-                System.out.println("ERRO. El diredtorio no existe");
+                System.out.println("ERROR. El diredtorio no existe");
             }
 
 
@@ -35,6 +35,8 @@ public class InterpreteSQL {
             createTAble(sentencia);
 
         } else if ((matcher = Pattern.compile("SELECT\\s+\\*\\s+FROM (.+);$", Pattern.CASE_INSENSITIVE).matcher(sentencia)).find()) {
+            String nombre = matcher.group(1);
+            String columnas = matcher.group(2);
             select(sentencia);
         } else if ((matcher = Pattern.compile("DROP TABLE\\s+(\\w+);$", Pattern.CASE_INSENSITIVE).matcher(sentencia)).find()) {
             dropTable(sentencia);
@@ -57,58 +59,87 @@ public class InterpreteSQL {
     }
     public boolean validar(String ruta) {
         File directorio = new File(ruta);
-            return directorio.exists() && directorio.isDirectory();
+        return directorio.exists() && directorio.isDirectory();
 
     }
     //........
     public void showTable(String mostrarTablas) {
+
 // PONERLE UN TRY CATCH ----------------
-        File directorio = new File(rutaTrabajo);
-        File[] archivos = directorio.listFiles();
-        csv.leerShow(archivos, rutaTrabajo);
+        if (rutaTrabajo != null){
+            File directorio = new File(rutaTrabajo);
+            File[] archivos = directorio.listFiles();
+            csv.leerShow(archivos, rutaTrabajo);
+        }else{
+            System.out.println("No se ha elegido una ruta de trabajo.");
+        }
 
     }
 
     public void createTAble(String tablaCreada) {
 
+        if(rutaTrabajo !=null){
+
+
         Matcher match;
         Pattern pattern = Pattern.compile("CREATE TABLE (\\w+) \\((.+)\\);", Pattern.CASE_INSENSITIVE);
         match = pattern.matcher(tablaCreada);
         if(match.find()){
-            String nombreTabla = match.group(2) + ".csv";
+            String nombreTabla = match.group(1);
             //System.out.println("Tabla " + nombreTabla);
-        String col = match.group(2);
+            String col = match.group(2);
             String[] columnas = col.split(",");
             // System.out.println("COlumnas:");
             csv.crearTabla(nombreTabla, columnas, rutaTrabajo);
         }else {
             System.out.println("Sintaxis con CREATE TABLE incorrecto");
         }
-
+        }else {
+            System.out.println("No se ha elegido la ruta de trabajo");
+        }
     }
 
     public void select(String selectTable) {
+        if (rutaTrabajo != null){
+
+        }else{
+            System.out.println("No se ha elegido la ruta de trabajo");
+        }
 
 
     }
 
     public void dropTable(String dropTable) {
+        if (rutaTrabajo != null){
         Matcher match;
         Pattern pattern = Pattern.compile("DROP TABLE (\\w+);$",Pattern.CASE_INSENSITIVE);
         match = pattern.matcher(dropTable);
         if (match.find()) {
             String nombreTabla = match.group(1) + ".csv";
-            System.out.println(match.group(1));
             File directorio = new File(rutaTrabajo + "/" + nombreTabla);
             if (directorio.exists()) {
+                try {
+
+
+                System.out.println("¿Realmente deseas borrar la tabla? s = si/n = no");
+                String res = bufer.readLine().trim().toLowerCase();
+
+                if (res.equals("s")){
                 if (directorio.delete()) {
-                    System.out.println("Ha sido borrado la tabla " + dropTable);
+                    System.out.println("Ha sido borrado la tabla " + match.group(1));
                 } else {
                     System.out.println("No se pudo borrar la tabla");
+                }}else if( res.equals("n")){
+                    System.out.println("Se ha cancelado la operacion");
+                }
+                } catch (IOException e){
+                    e.printStackTrace();
                 }
             } else {
                 System.out.println("La tabla no existe");
             }
+        }}else {
+            System.out.println("No se ha elegido la ruta de trabajo");
         }
     }
 
@@ -141,9 +172,4 @@ public class InterpreteSQL {
 
 
 
-
-
-
 }
-
-
